+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 Version 0.2.0.0
 ---------------
 
-*Apr 30, 2018*
+*May 1, 2018*
 
 <https://github.com/mstksg/backprop/releases/tag/v0.2.0.0>